Update README.md
Advanced GIS is a half-semester course at Carnegie Mellon University which builds upon previous GIS coursework. While focusing heavily on analyzing data and managing projects using ArcGIS and other Esri software, much of the coursework also includes experience using other GIS tools. This repository serves as a portfolio for much of my experience in geospatial analysis. It begins with a link to a final project submission for the preceding GIS course, which analyzes brain drain effects in Wyoming. It proceeds with various deliverables for the Advanced GIS course.